Users Controller: moved UserController /:usernameSlug to own dir
